Initialize number of counters to 0 before querying the capabilities (#3546)
Initialize the argument sai_stat_capability_list_t to a zero size when calling sai_query_stats_capability for the first time in initCounterCapabilities.

Why I did it

Currently, it is initialized as the sum of all the counter IDs that orchagent requires. This works, but it is misleading because the purpose of this SAI API is to return all the counters that are supported by the vendor SAI, regardless of what the SAI client requires.

Usually, it should be either 0 or the number of counters that vendor SAI supports. It is strange to pass any other value for vendor SAI and might indicate an error. Some vendors can report an error in this case.
To avoid this kind of error message, we always pass 0 in this case.