[DVS] Remove framework package during DVS build (#3577)
* remove and reinstall framework package during DVS build
What I did
When building the DVS image, remove the framework package when purging libswsscommon and other related packages. Re-install framework package when installing other packages.

Why I did it
DVS build is failing as swsscommon is a dependency of the framework package, so during the build the purge of libswsscommon fails. This package is not currently used for VS tests