Enable FDB learning event after all ports removed from default 1Q bridge (#3630)
What I did

This PR is to fix an orchagent crash issue. Error logs are as below.

2025 May  1 05:51:07.128331 str4-7060x6-64pe-7 ERR swss#orchagent: :- meta_generic_validation_remove: object 0x3a0000000000d7 reference count is 1, can't remove
2025 May  1 05:51:07.128331 str4-7060x6-64pe-7 ERR swss#orchagent: :- removeDefaultBridgePorts: Failed to remove bridge port, rv:-17
2025 May  1 05:51:07.128566 str4-7060x6-64pe-7 INFO swss#supervisord: orchagent terminate called after throwing an instance of 'std::runtime_error'
2025 May  1 05:51:07.128566 str4-7060x6-64pe-7 INFO swss#supervisord: orchagent   what():  PortsOrch initialization failure
2025 May  1 05:51:07.815330 str4-7060x6-64pe-7 INFO swss#supervisord 2025-05-01 05:51:07,814 WARN exited: orchagent (terminated by SIGABRT (core dumped); not expected)
It's because FDB is learnt on the default bridge, which increased reference count for bridge port and caused port removal failure.

The issue is addressed by not setting SAI_SWITCH_ATTR_FDB_EVENT_NOTIFY when creating switch, and enable it after all ports being removed from default bridge.