[fpmsyncd]Fixing fpmsyncd to handle routes without protocol (#3657)
What I did
When a route is programmed through swssconfig or through redis, there might be no protocol string. This would result in the response from orchagent to have emtpy protocol field. fpmsyncd would throw an exception which when handled through the try catch would exit fpmsyncd with exit code zero.

Why I did it

Handle empty protocol gracefully in fpmsyncd.
Remove the generic catch and exit 0 in fpmsyncd so that we will have a core dump and non zero exit on exceptions