[vs][mirror]: Update test to use the max TC number provided by VS lib (#3712)
The PR is intended to provide a fix for the change #1957

The old test behavior relies on a hardcoded max TC value of 255. Now it needs to be updated following the SAI_SWITCH_ATTR_QOS_MAX_NUMBER_OF_TRAFFIC_CLASSES attribute implementation in VS lib.

Due to the circular dependency, in order to have clean sonic-sairedis CI,
the current change should be merged first.