[Mux] Clear bulkers when rolling back mux switchover (#3788)
What I did
When a switchover failure is detected in MuxOrch, clear relevant bulkers to provide a clean slate for the rollback process.

Why I did it
In certain failure scenarios, if an exception is thrown inside the bulker, it's possible that the bulker is not cleared and still contains data in creating_entries or removing_entries. When the rollback process begins, these entries will be programmed to the SAI a second time, which is a) incorrect b) could potentially trigger the same exception second time.

How I verified it
Run the MuxRollbackTest.StandbyToActiveExceptionRollbackToStandby test