Avoid notifying counter polling for duplicated priority groups or queues (#3885)
* Avoid notifying counter polling for duplicated priority groups or queues
What I did

Avoid notifying counter polling for duplicated PG or queue when create_only_config_db_buffers is False in the flow:

Flex counter orchagent enables counter polling for PG or queue. It notifies counter polling for all PGs or queues since create_only_config_db_buffers is False
Particular items in BUFFER_PG or BUFFER_QUEUE is configured. It notifies counter polling for the items again.
In most cases, BUFFER_PG and BUFFER_QUEUE are handled ahead of counter polling (i.e., the order is 2, 1), and the issue won't occur. But in a rare scenario, it can be 1 and then 2.