[routeorch] Remove updateRoute call to prevent set operations (#3923)
This call to updateRoute is not needed because the route entry has already been removed. It is causing unneccesary route set operations and printing harmless (but verbose) error logs. Removing this call to prevent the side-effects.

What I did
Removed call to updateRoute from removeRoutePost path.

Why I did it
This call is not needed, and is causing error logs.